requirements.txt
openai==0.27.2
itchat==1.3.10
pillow==9.4.0
pyyaml==3.13
requests==2.28.2
func_timeout==4.3.5
